feat: export `VDataTable` types
